+++ app/controllers/admin/cms/locations_controller.rb
@@ -1,6 +1,8 @@
 class Admin::Cms::LocationsController < Admin::BaseController
   
   def index
+    authorize Location
+    
     @locations = Location.order('name')
     @locations = @locations.where("name LIKE '%#{params[:q]}%'") unless params[:q].nil?
     
@@ -16,12 +18,12 @@ class Admin::Cms::LocationsController < Admin::BaseController
   end
 
   def new
-    @location = Location.new name: 'New location'
+    @location = authorize Location.new(name: 'New location')
     render 'edit'
   end
 
   def create
-    @location = Location.new(location_params)
+    @location = authorize Location.new(location_params)
     
     if @location.save
       redirect_to action: 'show', id: @location.id, notice: 'Location was successfully created.'
@@ -31,15 +33,15 @@ class Admin::Cms::LocationsController < Admin::BaseController
   end
 
   def show
-    @location = Location.find(params[:id])
+    @location = authorize Location.find(params[:id])
   end
 
   def edit
-    @location = Location.find(params[:id])
+    @location = authorize Location.find(params[:id])
   end
 
   def update
-    @location = Location.find(params[:id])
+    @location = authorize Location.find(params[:id])
     
     if @location.update(location_params)
       redirect_to action: 'show', id: @location.id, notice: 'Location was successfully updated.'
@@ -49,7 +51,7 @@ class Admin::Cms::LocationsController < Admin::BaseController
   end
 
   def destroy
-    @location = Location.find(params[:id])
+    @location = authorize Location.find(params[:id])
     @location.destroy
     redirect_to action: 'index', notice: 'Location has been deleted.'
   end
@@ -57,13 +59,17 @@ class Admin::Cms::LocationsController < Admin::BaseController
   
   def pictures
     @location = Location.find(params[:id])
+    authorize  @location, :show?
   end
   
   def categories
     @location = Location.includes(:categories).find(params[:id])
+    authorize  @location, :show?
   end
   
   def create_categories
+    authorize Location, :update?
+    
     LocationCategory.delete_all location_id: params[:id]
     category_ids = params[:category_ids]
     
@@ -76,6 +82,7 @@ class Admin::Cms::LocationsController < Admin::BaseController
   
   def extra_properties
     @location = Location.find(params[:id])
+    authorize  @location, :show?
     5.times { @location.extra_properties.build }
   end
